\docType{data}
\name{mdat}
\alias{mdat}
\title{Cold stress (metabolites)}
\description{
  GC-MS dataset of galactose synthesis related metabolites
  during high CO2 stress in Arabidopsis thaliana.
}
\author{
  Henning Redestig
}
\references{
  Dutta, B., Kanani, H., Quackenbush, J., and Klapa, M. I.
  (2009). Time-series integrated omic analyses to elucidate
  short-term stress-induced responses in plant liquid
  cultures. Biotechnol Bioeng, 102(1), 264-279.
}
\keyword{datasets}

