\docType{data}
\name{tdat}
\alias{tdat}
\title{Cold stress (transcripts)}
\description{
  Microarray dataset of gene expression during severe cold
  stress in Arabidopsis thaliana. The dataset is an excerpt
  of galactose metabolism related genes and an additional
  990 random transcripts.
}
\author{
  Henning Redestig
}
\references{
  Dutta, B., Kanani, H., Quackenbush, J., and Klapa, M. I.
  (2009). Time-series integrated omic analyses to elucidate
  short-term stress-induced responses in plant liquid
  cultures. Biotechnol Bioeng, 102(1), 264-279.
}
\keyword{datasets}

