\name{trainmet}
\title{Metabolite-transcript co-responses}
\description{Calculate association statistics between metabolite and transcrips
using time-series data.}
\details{\tabular{ll}{
Package: \tab trainmet \cr
Type: \tab Package \cr
Version: \tab 0.0.1 \cr
Date: \tab 2011-03-23 \cr
Depends: \tab Biobase, pls \cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes \cr
}

A package for calculating association statistics between
metabolites and transcripts via classical correlation methods and
(via an external Python script) Hidden Markov Models.}
\alias{trainmet}
\docType{package}
\author{Henning Redestig \email{henning.red@gmail.com}}
