\name{mdat}
\title{Cold stress (metabolites)}
\usage{data(mdat)}
\description{GC-MS dataset of galactose synthesis related metabolites during
high CO2 stress in Arabidopsis thaliana.}
\alias{mdat}
\docType{data}
\keyword{datasets}
\references{Dutta, B., Kanani, H., Quackenbush, J., and Klapa,
M. I. (2009). Time-series integrated omic analyses to elucidate
short-term stress-induced responses in plant liquid
cultures. Biotechnol Bioeng, 102(1), 264-279.}
\author{Henning Redestig}
